feat(cluster-tax): close u64 cluster-wealth seam + real-curve M2 sim harness (#626 PR 3/3) (#629)
* feat(cluster-tax): close u64 cluster-wealth seam + real-curve M2 sim harness (#626 PR 3/3)

PR 3 of 3 for #626. Two parts:

A. Close the u64 seam (consensus-breaking; rides the protocol-4.0.0 reset).
   Widen the fee/factor/centroid APIs to u128 wealth end-to-end so the
   log-domain curve (PR1 #627) and the u128 accumulator (PR2 #628) are no
   longer throttled by a u64 ceiling:
   - FeeConfig::{cluster_factor, estimate_typical_fee, estimate_fee_*,
     compute_fee*} take u128 cluster_wealth.
   - ring_centroid_implied_factor(&[(u64, u128)]) — the C7 consensus
     fee-floor path (#602) in ledger/store.rs — takes u128 member wealth.
   - Removed PR2's `.min(u64::MAX as u128)` clamps at the three named sites:
     mempool relay fee (mempool.rs), C7 ring-centroid floor (store.rs), and
     the CLI estimate / RPC estimateFee paths (send.rs, rpc/mod.rs).
   - Re-derived the multiply bound each clamp was protecting with explicit
     comments: value(u64) × wealth(u128) can exceed u128 at astronomically
     distant cumulative wealth, so the centroid math uses saturating_mul /
     saturating_add, pinning overflow to u128::MAX → factor 6000 (max floor),
     the conservative anti-gaming direction and identical on every node (no
     fork). Deterministic, no floats.

B. Sim-harness gap closure (spec §7). New library harness
   cluster-tax/src/simulation/m2.rs exercises the REAL production
   ClusterFactorCurve: population declared in BTH, converted via PICO_PER_BTH
   at the curve boundary, per-agent factors derived from simulated cumulative
   cluster wealth through the same compiled factor() the node runs. The legacy
   hardcoded-factor entry point (add_owner_with_factor, the #314 harness) is
   kept but marked deprecated; new work uses add_owner_via_production_curve.
   Implements the #605 run matrix as CLI subcommands:
   - m2-cumulative (run set 1): 10/20yr horizons, honest + gamed, emits
     Δgini vs >0.05, merchant mispricing, whale progressivity.
   - m2-decay (run set 2): epoch-halving decay (half-lives 2/5/10yr, integer
     w >>= 1 per epoch — pure function of height, M3 lesson), plus wash-trading
     evasion (<20% gate) and privacy ring-ID rate (<50% gate) per ANALYSIS.md.
   Documented in --help and experiments/M2_RUNBOOK.md. Full suite is NOT run
   here (#605 owns the empirical runs); tiny-horizon --smoke variants prove
   each run type executes and emits its metrics as unit tests.

Resumes a partial build (previous builder hit its session limit). ZkFeeCurve
re-anchoring already landed in PR1 (#627). Gates: cargo test -p
bth-cluster-tax, cargo test -p botho --lib, cargo build --workspace,
cargo fmt --all -- --check — all green.


* fix(consensus): saturating_mul on C7 ring-centroid fee-floor wealth (#626 PR3)

Judge blocker: store.rs:2370 (ring_centroid_floored_factor, C7 consensus
fee-floor path) computed `entry.weight as u128 * global_wealth` with an
unchecked multiply. `global_wealth` is the unbounded, monotonic PR2 cluster-
wealth accumulator, so the product is not provably below u128::MAX: debug
validators panic (block-validation halt vector), release nodes wrap to a
WRONG, LOWER centroid -> LOWER fee floor (the non-conservative direction),
contradicting this PR's own saturation determinism claim.

- Wrap the product in `saturating_mul` so it saturates to u128::MAX before
  the divide; a saturated wealth maps deterministically to the max, most-
  conservative factor 6000 -> identical on every node -> no fork.
- Rewrite the adjacent comment: it previously implied the helper
  (ring_centroid_implied_factor) covered this product's overflow. It does
  not; it only saturates the internal value*member_wealth product. The
  comment now states accurately what saturates where.

Drive-by hardening of the sibling sites the Judge identified as the same
panic-capable `weight * global` pattern:
- compute_cluster_wealth_for_utxos (store.rs:~2436)
- the lottery-tilt wealth path (store.rs:~3006)
both now use saturating_mul.

Non-blocking nits:
- add_owner_with_factor: promote doc-only deprecation to the real
  #[deprecated] attribute; scope targeted #[allow(deprecated)] on the crate's
  own historical-reproduction call sites (the production-curve wrapper, the
  #314 regression tests, and run_lottery_experiment) so the build stays clean.
- fix stale `mc-cluster-tax` crate name in the cli-feature error message
  (-> bth-cluster-tax).

Regression test: test_ring_centroid_floor_saturates_at_max_wealth_no_panic
seeds cluster wealth = u128::MAX and asserts the C7 path does not panic
(debug would catch the old bug) and saturates to factor(u128::MAX).


---------

+++ transaction/core/src/validation/cluster_fee.rs
@@ -193,7 +193,7 @@ pub fn validate_cluster_fee(
     let minimum_fee = fee_config.compute_fee(
         TransactionType::Hidden,
         tx_size_bytes,
-        effective_wealth,
+        effective_wealth as u128,
         num_memos,
     );
 
@@ -238,7 +238,7 @@ pub fn validate_cluster_fee_dynamic(
     let minimum_fee = fee_config.compute_fee_with_dynamic_base(
         TransactionType::Hidden,
         tx_size_bytes,
-        effective_wealth,
+        effective_wealth as u128,
         num_memos,
         dynamic_base,
     );
@@ -275,7 +275,7 @@ pub fn compute_cluster_factor(
 ) -> u64 {
     let effective_wealth =
         compute_effective_cluster_wealth(input_tx_outs, input_values, cluster_wealth);
-    fee_config.cluster_factor(effective_wealth)
+    fee_config.cluster_factor(effective_wealth as u128)
 }
 
 /// Trait for looking up cluster wealth.
@@ -486,7 +486,7 @@ pub fn validate_ring_cluster_fee(
     let minimum_fee = fee_config.compute_fee_with_outputs(
         TransactionType::Hidden,
         tx_size_bytes,
-        max_cluster_wealth,
+        max_cluster_wealth as u128,
         num_outputs,
         num_memos,
     );
@@ -536,7 +536,7 @@ pub fn validate_ring_cluster_fee_dynamic(
     let minimum_fee = fee_config.compute_fee_with_dynamic_base_and_outputs(
         TransactionType::Hidden,
         tx_size_bytes,
-        max_cluster_wealth,
+        max_cluster_wealth as u128,
         num_outputs,
         num_memos,
         dynamic_base,
